[sdc] move some functions of `utils`-module to new module of `products`-subpackage
