+++ sdc/products/mswep.py
@@ -4,7 +4,7 @@ from odc.geo.xr import assign_crs
 from typing import Optional, Tuple
 from xarray import DataArray
 
-import sdc.utils as utils
+from sdc.products import _ancillary as anc
 from sdc.products import _query as query
 
 
@@ -39,7 +39,7 @@ def load_mswep(bounds: Tuple[float, float, float, float],
     The MSWEP data is available as daily precipitation estimates at 0.1° resolution.
     For more product details, see: https://www.gloh2o.org/mswep
     """
-    nc_files = query.filter_mswep_nc(directory=utils.get_catalog_path(product='mswep'),
+    nc_files = query.filter_mswep_nc(directory=anc.get_catalog_path(product='mswep'),
                                      time_range=time_range,
                                      time_pattern=time_pattern)
     ds_list = []
