+++ sdc/products/sanlc.py
@@ -4,7 +4,7 @@ from odc.stac import load as odc_stac_load
 from typing import Tuple
 from xarray import DataArray
 
-import sdc.utils as utils
+from sdc.products import _ancillary as anc
 from sdc.products import _query as query
 
 
@@ -29,10 +29,10 @@ def load_sanlc(bounds: Tuple[float, float, float, float]
     bands = ['nlc']
     
     # Load and filter STAC Items
-    catalog = Catalog.from_file(utils.get_catalog_path(product=product))
+    catalog = Catalog.from_file(anc.get_catalog_path(product=product))
     _, items = query.filter_stac_catalog(catalog=catalog, bbox=bounds)
     
-    common_params = utils.common_params()
+    common_params = anc.common_params()
     common_params['resampling'] = 'nearest'
     
     # Turn into dask-based xarray.Dataset
